requirements.txt
boto3==1.28.42
botocore==1.31.85
build==1.0.3
CacheControl==0.13.1
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
crashtest==0.4.1
distlib==0.3.8
dulwich==0.21.7
epl.geometry==1.0.3
epl.protobuf.v1==1.0.4
fastjsonschema==2.19.1
filelock==3.13.1
google-api-core==2.17.1
google-auth==2.28.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-tools==1.57.0
idna==3.6
importlib-metadata==7.0.1
installer==0.7.0
jaraco.classes==3.3.1
jmespath==1.0.1
keyring==24.3.0
more-itertools==10.2.0
msgpack==1.0.7
nsl.stac @ git+https://github.com/pantierra/stac-client-python.git@c7a29b13b4558d02e677bc2309feb2cdc32992cf
numpy==1.26.4
packaging==23.2
pexpect==4.9.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
protobuf==4.24.4
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyproject_hooks==1.0.0
python-dateutil==2.8.2
rapidfuzz==3.6.1
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.6.2
shapely==2.0.3
shellingham==1.5.4
six==1.16.0
tenacity==8.2.3
tomlkit==0.12.3
trove-classifiers==2024.1.31
urllib3==2.0.7
virtualenv==20.25.0
xattr==0.10.1
zipp==3.17.0
